start refactor to building the yaml inside dagster
